refactor: move lib/utils to types/xy_chart/utils
