+++ src/types/xy_chart/utils/dimensions.test.ts
@@ -1,10 +1,10 @@
-import { AxisTicksDimensions } from '../axes/axis_utils';
-import { AxisSpec, Position } from '../series/specs';
-import { LIGHT_THEME } from '../../utils/themes/light_theme';
-import { LegendStyle } from '../../utils/themes/theme';
+import { AxisTicksDimensions } from '../../../lib/axes/axis_utils';
+import { AxisSpec, Position } from '../../../lib/series/specs';
+import { LIGHT_THEME } from '../../../utils/themes/light_theme';
+import { LegendStyle } from '../../../utils/themes/theme';
 import { computeChartDimensions } from './dimensions';
-import { AxisId, getAxisId, getGroupId } from '../../utils/ids';
-import { Margins } from '../../utils/dimensions';
+import { AxisId, getAxisId, getGroupId } from '../../../utils/ids';
+import { Margins } from '../../../utils/dimensions';
 
 describe('Computed chart dimensions', () => {
   const parentDim = {
